feeat: like and dislike cards

+++ src/utils/api.js
@@ -84,25 +84,25 @@ class Api {
     .then((res) => this._checkResponse(res));
   }
 
-  setLike(data) {
-    return fetch(`${this._serverUrl}cards/${data}/likes`, {
-      method: 'PUT',
-      headers: {
-        authorization: this._token,
-        'Content-Type': 'application/json'
-      }
-    })
-    .then((res) => this._checkResponse(res));
-  }
-
-  deleteLike(data) {
-    return fetch(`${this._serverUrl}cards/${data}/likes`, {
-      method: 'DELETE',
-      headers: {
-        authorization: this._token
-      }
-    })
-    .then((res) => this._checkResponse(res));
+  changeLikeCardStatus(data, isLiked) {
+    if (!isLiked) {
+      return fetch(`${this._serverUrl}cards/${data}/likes`, {
+        method: 'PUT',
+        headers: {
+          authorization: this._token,
+          'Content-Type': 'application/json'
+        }
+      })
+      .then((res) => this._checkResponse(res));
+    } else {
+      return fetch(`${this._serverUrl}cards/${data}/likes`, {
+        method: 'DELETE',
+        headers: {
+          authorization: this._token
+        }
+      })
+      .then((res) => this._checkResponse(res));
+    }
   }
 
   getInitialData() {
